Create passwordrd.yml
This workflow automates the build and publication of the passwordrd package to PyPI whenever a new version tag (e.g., v1.0.0) is pushed to the repository. > > It leverages GitHub Actions and uses Trusted Publisher authentication to securely and reliably deploy Python distributions in line with best practices. > > Main steps include: > - Cloning the repository > - Setting up Python 3.11 > - Upgrading pip and installing the build module > - Building the package distribution > - Publishing to PyPI